https://trello.com/c/GVkcfL86/1-adding-translation-functions-for-data-objects
Refactored data structs to anticipate different values that may or may not be there from questions. Added translation layer for data to be converted to local values to be used. Created Utilities class for static methods that would be used across the app. Renamed VC to HomeVC, along with storyboard.